fixup! KVM: arm64: Instantiate VM shadow data from EL1
KVM: arm64: Initialize shadow vcpus with separate hypercalls

This is done as the first step towards donating memory per vcpu
in future patches without having to spend potentially too much
time in one hypercall.

Moreover, this has the nice effect of removing the need for
stashing the host vcpus in the memory donated for the pgd.

Signed-off-by: Fuad Tabba <tabba@google.com>