ANDROID: KVM: arm64: Refactor nvhe Makefile
Building and linking hypervisor modules into kernel modules involves very
similar problems to building and linking the KVM nVHE hypervisor object
into the kernel image. In order to re-use the same building procedure,
let's factor out and generalize the nVHE makefile in a way that allows it
to be re-used for module builds.

Signed-off-by: Quentin Perret <qperret@google.com>