ANDROID: arm64: insn: Add aarch64_addr_write
The process of applying hypervisor relocations involves patching
addresses in the hypervisor object. In the existing KVM nVHE relocation
procedure, the relocations are applied early enough for write-permission
not to be a problem when touching e.g. the rodata section. But applying
relocations on hypervisor modules embedded in a kernel module proves more
challenging, as the kernel module loader will actively map text and rodata
sections read-only. In order to allow patching in those sections, let's
introduce a new helper function using the text fixmap to temporarily map
the relevant pages writable.

Signed-off-by: Quentin Perret <qperret@google.com>