ANDROID: KVM: arm64: Be careful when removing MMIO stage-2 mappings
When the memory pool from which host stage-2 page-table pages are
allocated runs out of pages, we currently unilaterally unmap all
non-memory mappings to refill the pool. Indeed, as we don't support
device assignment yet, it is assumed that MMIO pages should all be owned
by the host in a default state.

In preparation for allowing to have non-default stage-2 permissions for
non-memory mappings, let's rework this procedure to check the state
of the MMIO page prior to unmapping it.

This sadly means that it is theoretically possible for the host stage-2
memory pool to run out pages without a way to re-fill it. This should
only happen if all of memory is mapped with page-granularity and we have
over 1GiB of non-memory pages that have been with mapped with a
non-default memory permission.

Signed-off-by: Quentin Perret <qperret@google.com>