KVM: arm64: Allow to pin shared memory owned by the hypervisor
The pKVM hypervisor currently can only pin shared memory owned by the
host, which makes sense as only the only sharing direction is host to
hypervisor. But this will change... In preparation to allow the host the
donate memory to the hyp but having it shared back, readonly (or
wrprotect).

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>