KVM: arm64: Extend memory sharing to allow wrprotected donations
Sharing memory between the hypervisor and the host where the latter only
retains read-only access is interesting for some features such as
tracing to reduce attack surface.

Reducing the permissions on the host own memory implies an ownership
transfer. To avoid a costly host-hyp donation followed by a hyp-host
sharing, brings a new donation-sharing mechanism. This is similar to the
regular sharing but the ownership is also transfered and the initiator
can set reduce permissions.

For now this mechanism is limited to hyp/host sharing with wrprotect as
this is what we're interested in. The wrprotect value is cached into the
hyp vmemmap to avoid having to walk the table for later permission
verifications.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>